POC for extract-macros
Added `extract-macros`, which works exactly like `import-macros`, but it
targets a normal fennel module and returns a table of macros that were
set on top-level module scope.

This doesn't allow explicit exporting of macros - any macros on the top
module scope will show up in the macro table, which in terms of design
goals likely falls show of how explict we want this to be.

The other problem with this appraoch is that if one has both
`my-module/init.fnl` and `my-module/init-macros.fnl`, both would resolve
to `macro-loaded.my-module`, creating a naming conflict. This patch
solves this by adding an `extract:` prefix when caching the macros in
`fennel.macro-loaded`, but there may be cleaner solutions worth
considering.