re-implement pick-values as a special
Doing so introduces a number of optimizations and improvements:

For (pick-values 1 ...), instead of generating syms to assign and
truncate the return values accordingly, we just compile the target
expression(s) and wrap the result in ( <ret> ), preserving side effects
in the process. For example, `(pick-values 1 (foo))` emits `(foo())`

For n = 0, we simply preserve side effects; the result will be either no
values or nil depending on outer context. `[(pick-values 0 ...)]` will
be `{}`, but `[x (pick-values 0 :y) :z]` will result in
`{'x', nil, 'z'}}`.

For all n > 1, instead of the `let` form with `n` bindings in the macro
version, we simply use destructure to generate the declaration +
assignment of n syms. This reduces the number of unnecessary IIFE's even
for n > 1.

---

BUG FIXES

Previously, `(pick-values 0 ...)` would erroneously drop the side
effects from the expressions passed to the macro. This new version
correctly preserves them.

---

NOTES

As with `(values)`, `pick-values` for n > 1 does not guarantee
order of evaluation for its operand expressions (relevant for side
effects). Multi-value assignment in Lua leaves order unspecified as
well, but this can still come as a surprise.