adjust min propagated `nval` in `values`
This commit is here to show what would happen when changing the minimum
`nvals` in `values` when adjusting for operand position. In
dbb10b59250a1dbcd2b7a86d0ef5ae0e538b0388, the minimum was set to 1, but
strictly speaking it should be 0.

The main impact of this is that it changes at which point the compiler
treats further expressions (lists only) in `values` as a side effect,
and pulls them out of a `(local (x y) (values <expr> <expr> ...))`
assignment to run first.

To see the impact, see the changes in test/core.fnl